Death circle deals damage

See merge request null-ptr/projet-synthese!53

+++ Assets/Scripts/Playmode/Event/PlayerHealthEvent.cs
@@ -5,10 +5,10 @@ namespace ProjetSynthese
     public class PlayerHealthEvent : IEvent
     {
         public Health PlayerHealth { get; private set; }
-        public int OldHealthPoints { get; private set; }
-        public int NewHealthPoints { get; private set; }
+        public float OldHealthPoints { get; private set; }
+        public float NewHealthPoints { get; private set; }
 
-        public PlayerHealthEvent(Health playerHealth, int oldHealthPoints, int newHealthPoints)
+        public PlayerHealthEvent(Health playerHealth, float oldHealthPoints, float newHealthPoints)
         {
             PlayerHealth = playerHealth;
             OldHealthPoints = oldHealthPoints;
